feat: add ProofDeliveryStatusView component
